feat:  start of the scene and make enemy inherit from entity
